Configure Apollo client; hook up Hasura database
I add a dummy component `<DisplayUsers/>` that fetches user data from the
hasura server and puts them in a `ScrollView`. Currently the the DB just
contains 3 dummy users.

The database address is hardcoded to the *local IP-address
of my laptop*. In the future, contributors ought to put their own
database's IP-address in an environment variable, which git should
ignore, and Apollo Client should read that address.